refactor: collapse Accounts' twin event factories into AccountsEvents
Neither twin's name survives: 5 of the 7 contexts already name their factory
`<Context>Events`, and `UserEvents` could not hold `staff_invitation_sent`
honestly. The merged module keeps both call shapes because both are real —
the user factories take the user (so they can validate it and derive the
payload), the staff factories take an id (their producers hold one and not a
user).

Unlike ProgramCatalog's, these twins were not identical: the domain factory
validated a user struct and built a rich payload, the integration one took an
id and added `user_id`. The merge keeps both halves, with `user_id` put last
so caller payload can override the derived fields but never the identity
consumers key on.

`user_email_changed` goes. It was built, dispatched to a bus with no
handlers, and consumed by nothing — its `previous_email` "audit trail" was
never recorded anywhere. Same class as the four dead broadcasts removed in
PR 4. Its call site's `case` then had two identical branches, so it
collapses to the call it was wrapping.
