+++ test/klass_hero/messaging/adapters/driving/events/messaging_event_handler_test.exs
@@ -8,7 +8,7 @@ defmodule KlassHero.Messaging.Adapters.Driving.Events.MessagingEventHandlerTest 
   import KlassHero.EventTestHelper
   import KlassHero.Factory
 
-  alias KlassHero.Accounts.Domain.Events.AccountsIntegrationEvents
+  alias KlassHero.Accounts.Domain.Events.AccountsEvents
   alias KlassHero.AccountsFixtures
   alias KlassHero.Messaging.Adapters.Driving.Events.MessagingEventHandler
   alias KlassHero.Messaging.Message
@@ -37,9 +37,9 @@ defmodule KlassHero.Messaging.Adapters.Driving.Events.MessagingEventHandlerTest 
       )
 
       event =
-        AccountsIntegrationEvents.user_anonymized(
-          user.id,
-          %{anonymized_email: "deleted_#{user.id}@anonymized.local"}
+        AccountsEvents.user_anonymized(
+          %{id: user.id, email: "deleted_#{user.id}@anonymized.local"},
+          %{previous_email: user.email}
         )
 
       assert :ok == MessagingEventHandler.handle_event(event)
@@ -69,9 +69,9 @@ defmodule KlassHero.Messaging.Adapters.Driving.Events.MessagingEventHandlerTest 
       user = AccountsFixtures.user_fixture()
 
       event =
-        AccountsIntegrationEvents.user_anonymized(
-          user.id,
-          %{anonymized_email: "deleted_#{user.id}@anonymized.local"}
+        AccountsEvents.user_anonymized(
+          %{id: user.id, email: "deleted_#{user.id}@anonymized.local"},
+          %{previous_email: user.email}
         )
 
       assert :ok == MessagingEventHandler.handle_event(event)
@@ -84,9 +84,9 @@ defmodule KlassHero.Messaging.Adapters.Driving.Events.MessagingEventHandlerTest 
       user = AccountsFixtures.user_fixture()
 
       event =
-        AccountsIntegrationEvents.user_anonymized(
-          user.id,
-          %{anonymized_email: "deleted_#{user.id}@anonymized.local"}
+        AccountsEvents.user_anonymized(
+          %{id: user.id, email: "deleted_#{user.id}@anonymized.local"},
+          %{previous_email: user.email}
         )
 
       assert :ok == MessagingEventHandler.handle_event(event)
